# Speed improvements to resize kernel (w/ SIMD)