### Prerequisites

- [X] I have written a descriptive pull-request title
- [X] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [X] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [X] I have provided test coverage for my change (where applicable)

### Description
Related to #1476, this PR includes some speed improvements to the resize kernels.
In particular, it introduces a new vectorized path using AVX2/FMA operations to perform the convolutions.
For those interested, [here is](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABABgAJiBGAOgDkBXfGKASzFwG4BYAKDMq0ASgwB2GVsxoBhCPgAOrADYsAyiwBu7GFz4DqNEeMkwaASXFtRudrv4UDRiVIsYrNjjQAaADiQ9ePlZxFlFsJXJYbAATCFElAE9yMVxsADMYclw3BjAMciEdVgAvGABpUJglPgBvPnIG8nk2DWwMTKjY+KS0pQg2gCpyYAY0jKgABTcAxvJ6xuIAZnJg/IAZGFEAcwwAC3mGut5Z2YBtAFkYPYhoswUlAApL69v7gHl5CTjcGgBBLa2sFwNg0MAsSmCwS2AEoALoHE5bK4zRoAXwRCKW5AAajA8tAUORZKINBAlKDZLAHrA0ji8RgCZEIAB3VQYbBQDCFNLQhFHE6NXr9DBDEZjNTsznkAC85D2rB+YvGUygKIFQsGw1G4wAoqJojKteKoGyOfkANTkB7yn4bbZ7cgAMnIAD9FtC1SdcfioAAmACsSAAPBqMAA+SI6BhKDABpAAfQosu9DL9gZDfTaYZoAC0WBBPbMU9A4xnhRGgdHY4H41RDcW08HQ9m81ACwjZq0oOR8HgANb1+klwMyKLtB6kNDkSfTqczqhThfkJdUD18DuNZm7ZSZB5KiVm8hBo26/W844C8j8y+zSsxuOJw0AMV7NHOVdY8kSv2i0QeG5vBoAFVrHSUxflwIMG3QOkfVLZsw2pGBaTbVlJS5ZDoTQADAN+DQEF9GgJhYfAGHabEOR8BBFl9B4GzjUcYDaGBVDAcIORA1IMgeAYHliBhgBUAZoX3E10OhP5cFUKEVAeLCe37LCcJvO9qwTUg10CC9AMjXAqwfOtZRfbA3w/L8Eh/P9lMvTiwMkqChygGD6PTBCkNpWyMj+X93KZNCzW5KdfWhJTtJ0xo8IIoiSLImAKKgKiaLoxyGOkMcWLYpQONA7jeP4wSYGEvdtQPKVLWCiSIOk7ZZPk3tcD7ULwpOVSDM06yGlE00ytlFBC0vVDuow2lZRpchPPAnyxsG9DAvIFBNJvdEtJUqN7xrJNyEi7y/1ajapz2hNVwCHDWFpB5rW3W1Nh2XYnXId1yDDWVgpw68dMOx8jNfd8Y0/b8fI6k4Jvs6CpxcptM3DRDppZIbuSa5qIvwwjiKgUjyMo6jaIhxjmNY9ioAmni+IgAShJEkqxLNSqpJkmA5KnerGuwsLms+jSTrZy8uvQ8hzRe/qBRmgLkMNMaQcs3yRc5ObgqF8hlpwhsqF9HwyyzXSq1VnxDVUXBJt2ta1MTGgAHErjWFkWEZrX1vU82riA+R5BtkKuZvM6rUuhUaDtW77tXcgAEJZQ0t6gdUnXnx+syAas7mdMlyCwdg1MdY16Hpbh2bMNZpHGhVtW8fHAZeZp/OC7tjAdcWy8lcTygAHZyF46DhMdKO1c9ZblqAA=) a sharplab link with the current codegen for the `ConvolveCore` method, when using SIMD operations.

Still running benchmarks and work in progress...