@saucecontrol 2021-01-20
Looking good on Skylake 👍 

``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.102
  [Host]     : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
  Job-XNQYOQ : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT
  Job-ALEINB : .NET Core 2.1.24 (CoreCLR 4.6.29518.01, CoreFX 4.6.29518.01), X64 RyuJIT
  Job-LVTEAU : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
```

Before:

|                                   Method |        Job |       Runtime |      Mean |     Error |    StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------- |----------- |-------------- |----------:|----------:|----------:|------:|------:|------:|------:|----------:|
|                            SystemDrawing | Job-ILLXJH |    .NET 4.7.2 | 17.320 ms | 0.0918 ms | 0.0814 ms |  1.00 |     - |     - |     - |     256 B |
| &#39;ImageSharp, MaxDegreeOfParallelism = 1&#39; | Job-ILLXJH |    .NET 4.7.2 |  9.572 ms | 0.0701 ms | 0.0656 ms |  0.55 |     - |     - |     - |   40624 B |
|                                          |            |               |           |           |           |       |       |       |       |           |
|                            SystemDrawing | Job-MQEJZE | .NET Core 2.1 | 17.535 ms | 0.2673 ms | 0.2500 ms |  1.00 |     - |     - |     - |      96 B |
| &#39;ImageSharp, MaxDegreeOfParallelism = 1&#39; | Job-MQEJZE | .NET Core 2.1 |  8.645 ms | 0.0660 ms | 0.0618 ms |  0.49 |     - |     - |     - |   40472 B |
|                                          |            |               |           |           |           |       |       |       |       |           |
|                            SystemDrawing | Job-BXMGKS | .NET Core 3.1 | 17.498 ms | 0.1888 ms | 0.1766 ms |  1.00 |     - |     - |     - |     140 B |
| &#39;ImageSharp, MaxDegreeOfParallelism = 1&#39; | Job-BXMGKS | .NET Core 3.1 |  8.008 ms | 0.0335 ms | 0.0314 ms |  0.46 |     - |     - |     - |   40472 B |

After:

|                                   Method |        Job |       Runtime |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------- |----------- |-------------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|                            SystemDrawing | Job-KIBPBM |    .NET 4.7.2 | 17.318 ms | 0.1773 ms | 0.1572 ms |  1.00 |      - |     - |     - |     256 B |
| &#39;ImageSharp, MaxDegreeOfParallelism = 1&#39; | Job-KIBPBM |    .NET 4.7.2 |  9.562 ms | 0.0685 ms | 0.0607 ms |  0.55 |      - |     - |     - |   40624 B |
|                                          |            |               |           |           |           |       |        |       |       |           |
|                            SystemDrawing | Job-AOLXLZ | .NET Core 2.1 | 17.362 ms | 0.2023 ms | 0.1892 ms |  1.00 |      - |     - |     - |      96 B |
| &#39;ImageSharp, MaxDegreeOfParallelism = 1&#39; | Job-AOLXLZ | .NET Core 2.1 |  8.621 ms | 0.0707 ms | 0.0662 ms |  0.50 |      - |     - |     - |   40472 B |
|                                          |            |               |           |           |           |       |        |       |       |           |
|                            SystemDrawing | Job-INXCIO | .NET Core 3.1 | 17.418 ms | 0.1830 ms | 0.1712 ms |  1.00 |      - |     - |     - |     140 B |
| &#39;ImageSharp, MaxDegreeOfParallelism = 1&#39; | Job-INXCIO | .NET Core 3.1 |  6.545 ms | 0.0714 ms | 0.0668 ms |  0.38 | 7.8125 |     - |     - |   40424 B |