@Sergio0694 2021-01-20
@saucecontrol That's awesome, thank you for running the benchmarks on your machine! 🚀

Here are mine, it looks like this PR is actually ever so slightly slower than master.
I'm thinking it might just be that my Ryzen 2700X (Zen+ arch) is trash at FMA stuff, wouldn't surprise me 🤔

## Master

|                                   Method |      Runtime |      Mean |     Error |    StdDev | Ratio |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------- |------------- |----------:|----------:|----------:|------:|--------:|------:|------:|----------:|
| ImageSharp | .NET Core 3.1 |  7.506 ms | 0.0110 ms | 0.0098 ms |  0.44 |  7.8125 |     - |     - |   40772 B |

## PR

|                                   Method |      Runtime |      Mean |     Error |    StdDev | Ratio |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------------------- |------------- |----------:|----------:|----------:|------:|--------:|------:|------:|----------:|
| ImageSharp | .NET Core 3.1 |  7.646 ms | 0.0986 ms | 0.1174 ms |  0.45 |  7.8125 |     - |     - |   40712 B |