@antonfirsov 2021-01-20
According to the test failures, there is some noticeable (but visually still insignificant) difference between the `Vector4` and the FMA output. @Sergio0694 I wonder if we get a better understanding on this before changing tolerances?