@antonfirsov 2021-01-20
Regarding my https://github.com/SixLabors/ImageSharp/pull/1513#issuecomment-763597089, I think Wikipedia answered my concerns:

>  a fused multiply–add would compute the entire expression a+b×c to its full precision before rounding the final result down to N significant bits.

If I get it right, the FMA code is actually more accurate.