@Sergio0694 2021-01-20
Yeah, in theory the FMA path should be more accurate. From what @tannergooding said on Discord:

> "So, for reference, an fma(x, y, z) vs x * y + z should differ by no more than a single bit (ignoring the special case around infinity)
this can compound over many operations to produce even more significant differences
this falls out because the former rounds once while the latter rounds twice"

So in theory the difference should mean the FMA is ever so slightly better, as you mentioned.
It's not 100% that in theory since if the number of pixels per row is not evenly divisible, the last pixel will be computed with a slightly different precision compared to the previous, and also there's the fact that we're using 4 different accumulators in parallel and then just summing them all together at the end, so I'd expect that would alter the overall precision a bit too. But all things considered this new version should be at least equally as precise as the original, I think 🤔
Also if we can't spot any differences when looking at the images, I'd say we're probably good 😄