@saucecontrol 2021-01-20
> using 4 different accumulators in parallel and then just summing them all together at the end, so I'd expect that would alter the overall precision a bit too

Yes, this also improves precision.  Take the following example:

```c#
var rand = new Random(42);
var vals = new float[128];
for (int i = 0; i < vals.Length; i++)
	vals[i] = (float)rand.NextDouble();

float fa = 0;
for (int i = 0; i < vals.Length; i++)
	fa += vals[i];

double da = 0;
for (int i = 0; i < vals.Length; i++)
	da += vals[i];

float a0 = 0, a1 = 0, a2 = 0, a3 = 0;
for (int i = 0; i < vals.Length / 4; i++)
{
	int j = i * 4;
	a0 += vals[j];
	a1 += vals[j + 1];
	a2 += vals[j + 2];
	a3 += vals[j + 3];
}
float pa = a0 + a1 + a2 + a3;

Console.WriteLine("float accumulator:    " + fa);
Console.WriteLine("double accumulator:   " + da);
Console.WriteLine("4x float accumulator: " + pa);
```

Outputs:

```
float accumulator:    54.136913
double accumulator:   54.136910965666175
4x float accumulator: 54.13691
```